+++ src/auth.js
@@ -59,7 +59,6 @@ function makeURL(host, path, queryStringObj) {
 
 module.exports = {
     generateAuth: function (request, clientToken, clientSecret, accessToken, host, maxBody, guid, timestamp) {
-        maxBody = maxBody || 131072;
         guid = guid || uuid.v4();
         timestamp = timestamp || helpers.createTimestamp();